Final_Course_Project/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning and NLP
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# HTTP requests for local LLM
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For better performance with sentence transformers
scikit-learn>=1.3.0